refactor: rename stories to separate utils from components
